introducing Gate function

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
 // use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostsController extends Controller
 {
@@ -92,6 +93,14 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post!");
+        // };
+
+        $this->authorize('update-post',$post);
+
         return view('posts.edit', ['post'=>BlogPost::findOrFail($id)]);
     }
 
@@ -105,6 +114,12 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post!");
+        // };
+        $this->authorize('update-post',$post);
+
         $validated = $request->validated();
 
         $post->fill($validated);
@@ -125,6 +140,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog post!");
+        // };
+
+        $this->authorize('delete-post',$post);
+
         $post->delete();
 
         session()->flash('status','Blog post was deleted!');
